Abort config if libicu is not found.
The configuration script now produces a hard error if libicu is missing on the system since the build process can't complete without it, it checks both pkg-config and available headers to see if ICU is installed. A warning is show if icu is only partially detected, and the configure script errors out if ICU is not detected at all.