BREAKING CHANGE: LTS Rules use UTF-8 v2
This commit changes the format of LTS rules and therefore breaks
the CMU LTS rules in mimic-english.

This commit simplifies the structure of the previous one, and makes
it more cache-friendly by using an Array of Structs instead of an
Struct of Arrays in the LTS rules.

Before UTF-8 support each rule used 6 bytes.
In the previous commit, we were using 1+4+2+2 = 9 bytes per rule
This commit changes that so we use only 4+2+2 = 8 bytes (which is
much better to prevent structure alignment issues)