Define a public API unstable draft
This commit sets the default visibility to hidden and exports
all the symbols required to compile mimic-full.

Some of those symbols should be private, for instance:
 - ffeature_string should be private as there is the equivalent mimic_ffeature_string

This API should little by little be reduced to something well designed
and all public functions should be prefixed with mimic_ whenever possible
to prevent clashes.